requirements.txt
neo
numpy
pynwb
open-ephys-python-tools
holoviews
bokeh
streamz
matplotlib
